Change video recording quality to VGA and update preview layer accordingly.
The video displayed on the screen represents the final product (barring
some clipping on the sides on iPhone), so lowering the recording
quality affects the image to some degree. However, it is necessary to
lower the recording quality somewhat in order to not upload huge files
to the server. In the future, some post-recording compression may be
the correct choice; however, I don’t have time to implement that right
now.